chore: update WhatsNew content
- Introduce reusable gradient brush and apply to buttons and links for a modernized UI.
- Expand and update "What's New" section with new features and improvements.
- Add prominent "Join the Testnet" and "Download Preview" actions.
- Move Twitter post link logic to base ViewModel for reuse.
- Show updated WhatsNew dialog on next launch to highlight changes.
